changing how processes work
