+++ cmd/scollector/collectors/processes_darwin.go
@@ -2,6 +2,9 @@ package collectors
 
 import "fmt"
 
-func WatchProcesses(procs []*WatchedProc) error {
+func AddProcessConfig(line string) error {
 	return fmt.Errorf("process watching not implemented on Darwin")
+}
+
+func WatchProcesses() {
 }
