+++ cmd/scollector/main.go
@@ -51,8 +51,6 @@ var (
 	flagFreq            = flag.String("freq", "15", "Set the default frequency in seconds for most collectors.")
 	flagConf            = flag.String("conf", "", "Location of configuration file. Defaults to scollector.conf in directory of the scollector executable.")
 
-	procs []*collectors.WatchedProc
-
 	mains []func()
 )
 
@@ -115,11 +113,9 @@ func readConf() {
 		case "freq":
 			f(flagFreq)
 		case "process":
-			p, err := collectors.NewWatchedProc(v)
-			if err != nil {
+			if err := collectors.AddProcessConfig(v); err != nil {
 				slog.Fatal(err)
 			}
-			procs = append(procs, p)
 		case "keepalived_community":
 			collectors.KeepAliveCommunity = v
 		default:
@@ -193,11 +189,9 @@ func main() {
 			collectors.Vsphere(user, pwd, host)
 		}
 	}
-	if len(procs) > 0 {
-		if err := collectors.WatchProcesses(procs); err != nil {
-			log.Fatal(err)
-		}
-	}
+	// Add all process collectors. This is platform specific.
+	collectors.WatchProcesses()
+
 	if *flagFake > 0 {
 		collectors.InitFake(*flagFake)
 	}
